Better handle lyrics modifications.
When a track is updated with lyrics without any useful
content, we will now smartly handle it.

If the content is empty and there's no old version of
lyrics associated to the track, we'll ignore the change
entirely.

If the content is not empty and there is no old version
of lyrics, this counts as a "create" operation.

If the content is empty and there is an old version of
lyrics, this counts as a "delete" operation.